requirements.txt
# Core dependencies
# (no external core dependencies for base simulation)

# Testing
pytest==8.4.2
pytest-cov==6.0.0

# Development
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.13.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.0

# API (Unity Integration)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
# websockets==12.0  # Para futuro (WebSocket)

# Utilities
python-dotenv==1.0.1
PyYAML==6.0.1

# Voxel System
numpy>=1.24.0
Pillow>=10.0.0

pluggy==1.6.0
Pygments==2.19.2
tomli==2.3.0
typing_extensions==4.15.0


